setup creation of new style files and moved to separate predefined file
